moving veteran model to veteran::base
